requirements.txt
opencv-python == 4.7.0.72
numpy == 1.24.2
matplotlib == 3.7.1
pandas == 2.0.0

pydot == 1.4.2
# trained on cpu
tensorflow-cpu == 2.12.0
# tensorflow